+++ test/subscriptions/single_stream_subscription_test.exs
@@ -15,36 +15,48 @@ defmodule EventStore.Subscriptions.SingleStreamSubscriptionTest do
     {:ok, %{conn: conn}}
   end
 
-  test "create subscription to a single stream" do
-    stream_uuid = UUID.uuid4
-    {:ok, stream} = Streams.open_stream(stream_uuid)
+  describe "subscribe to stream" do
+    test "create subscription to a single stream" do
+      stream_uuid = UUID.uuid4()
+      {:ok, stream} = Streams.open_stream(stream_uuid)
 
-    subscription = create_subscription(stream_uuid, stream)
+      subscription = create_subscription(stream_uuid, stream)
 
-    assert subscription.state == :request_catch_up
-    assert subscription.data.subscription_name == @subscription_name
-    assert subscription.data.subscriber == self()
-    assert subscription.data.last_seen == 0
-    assert subscription.data.last_ack == 0
-  end
+      assert subscription.state == :request_catch_up
+      assert subscription.data.subscription_name == @subscription_name
+      assert subscription.data.subscriber == self()
+      assert subscription.data.last_seen == 0
+      assert subscription.data.last_ack == 0
+    end
 
-  test "create subscription to a single stream from starting stream version" do
-    stream_uuid = UUID.uuid4
-    {:ok, stream} = Streams.open_stream(stream_uuid)
+    test "create subscription to a single stream from starting stream version" do
+      stream_uuid = UUID.uuid4()
+      {:ok, stream} = Streams.open_stream(stream_uuid)
 
-    subscription = create_subscription(stream_uuid, stream, start_from_stream_version: 2)
+      subscription = create_subscription(stream_uuid, stream, start_from_stream_version: 2)
 
-    assert subscription.state == :request_catch_up
-    assert subscription.data.subscription_name == @subscription_name
-    assert subscription.data.subscriber == self()
-    assert subscription.data.last_seen == 2
-    assert subscription.data.last_ack == 2
+      assert subscription.state == :request_catch_up
+      assert subscription.data.subscription_name == @subscription_name
+      assert subscription.data.subscriber == self()
+      assert subscription.data.last_seen == 2
+      assert subscription.data.last_ack == 2
+    end
+
+    test "create subscription to a single stream with event mapping function" do
+      stream_uuid = UUID.uuid4()
+      {:ok, stream} = Streams.open_stream(stream_uuid)
+
+      mapper = fn event -> event.event_id end
+      subscription = create_subscription(stream_uuid, stream, mapper: mapper)
+
+      assert subscription.data.mapper == mapper
+    end
   end
 
   describe "catch-up subscription" do
     test "catch-up subscription, no persisted events" do
       self = self()
-      stream_uuid = UUID.uuid4
+      stream_uuid = UUID.uuid4()
       {:ok, stream} = Streams.open_stream(stream_uuid)
 
       subscription =
@@ -60,7 +72,7 @@ defmodule EventStore.Subscriptions.SingleStreamSubscriptionTest do
 
     test "catch-up subscription, unseen persisted events", %{conn: conn} do
       self = self()
-      stream_uuid = UUID.uuid4
+      stream_uuid = UUID.uuid4()
       {:ok, stream_id} = Stream.create_stream(conn, stream_uuid)
       recorded_events = EventFactory.create_recorded_events(3, stream_id)
       {:ok, 3} = Appender.append(conn, stream_id, recorded_events)
@@ -75,7 +87,7 @@ defmodule EventStore.Subscriptions.SingleStreamSubscriptionTest do
       assert subscription.state == :catching_up
       assert subscription.data.last_seen == 0
 
-      assert_receive {:events, received_events, ^self}
+      assert_receive {:events, received_events}
       assert_receive {:caught_up, 3}
 
       expected_events = EventFactory.deserialize_events(recorded_events)
@@ -115,7 +127,6 @@ defmodule EventStore.Subscriptions.SingleStreamSubscriptionTest do
   end
 
   test "notify events" do
-    self = self()
     stream_uuid = UUID.uuid4
     {:ok, stream} = Streams.open_stream(stream_uuid)
 
@@ -129,7 +140,7 @@ defmodule EventStore.Subscriptions.SingleStreamSubscriptionTest do
 
     assert subscription.state == :subscribed
 
-    assert_receive {:events, received_events, ^self}
+    assert_receive {:events, received_events}
 
     assert pluck(received_events, :correlation_id) == pluck(events, :correlation_id)
     assert pluck(received_events, :causation_id) == pluck(events, :causation_id)
@@ -155,7 +166,7 @@ defmodule EventStore.Subscriptions.SingleStreamSubscriptionTest do
 
       assert subscription.state == :subscribed
 
-      assert_receive {:events, received_events, ^self}
+      assert_receive {:events, received_events}
       assert length(received_events) == 3
 
       subscription =
@@ -177,7 +188,7 @@ defmodule EventStore.Subscriptions.SingleStreamSubscriptionTest do
       assert_receive {:caught_up, last_seen}
 
       # should not receive already seen events
-      refute_receive {:events, _received_events, ^self}
+      refute_receive {:events, _received_events}
 
       subscription = StreamSubscription.caught_up(subscription, last_seen)
 
@@ -197,7 +208,7 @@ defmodule EventStore.Subscriptions.SingleStreamSubscriptionTest do
       assert subscription.state == :catching_up
 
       # should receive already seen, but not ack'd, events
-      assert_receive {:events, received_events, ^self}
+      assert_receive {:events, received_events}
       assert length(received_events) == 3
 
       assert_receive {:caught_up, last_seen}
@@ -230,7 +241,7 @@ defmodule EventStore.Subscriptions.SingleStreamSubscriptionTest do
 
     assert subscription.state == :catching_up
 
-    assert_receive {:events, received_events, ^self}
+    assert_receive {:events, received_events}
     assert length(received_events) == 3
 
     assert_receive {:caught_up, 3}
@@ -245,7 +256,6 @@ defmodule EventStore.Subscriptions.SingleStreamSubscriptionTest do
   end
 
   test "should not notify events until ack received" do
-    self = self()
     stream_uuid = UUID.uuid4
     events = EventFactory.create_recorded_events(6, 1)
     initial_events = Enum.take(events, 3)
@@ -265,8 +275,8 @@ defmodule EventStore.Subscriptions.SingleStreamSubscriptionTest do
     assert subscription.data.last_ack == 0
 
     # only receive initial events
-    assert_receive {:events, received_events, ^self}
-    refute_receive {:events, _received_events, ^self}
+    assert_receive {:events, received_events}
+    refute_receive {:events, _received_events}
 
     assert length(received_events) == 3
     assert pluck(received_events, :correlation_id) == pluck(initial_events, :correlation_id)
@@ -282,7 +292,7 @@ defmodule EventStore.Subscriptions.SingleStreamSubscriptionTest do
     assert subscription.data.last_ack == 3
 
    # now receive all remaining events
-   assert_receive {:events, received_events, ^self}
+   assert_receive {:events, received_events}
 
    assert length(received_events) == 3
    assert pluck(received_events, :correlation_id) == pluck(remaining_events, :correlation_id)
@@ -291,8 +301,8 @@ defmodule EventStore.Subscriptions.SingleStreamSubscriptionTest do
   end
 
   defp create_subscription(stream_uuid, stream, opts \\ []) do
-    StreamSubscription.new
-    |> StreamSubscription.subscribe(stream_uuid, stream, @subscription_name, self(), self(), opts)
+    StreamSubscription.new()
+    |> StreamSubscription.subscribe(stream_uuid, stream, @subscription_name, self(), opts)
   end
 
   defp pluck(enumerable, field) do